issue/2890 Refactored for clarity, added _isInherited, ES6 modules
