+++ bower.json
@@ -1,7 +1,7 @@
 {
   "name": "adapt-contrib-trickle",
   "version": "5.0.0",
-  "framework": ">=5.7",
+  "framework": ">=5.8",
   "homepage": "https://github.com/adaptlearning/adapt-contrib-trickle",
   "issues": "https://github.com/adaptlearning/adapt_framework/issues/new",
   "extension" : "trickle",
