+++ example.json
@@ -7,6 +7,7 @@
     // Manual Configuration Options
     "_trickle": {
         "_isEnabled": true,
+        "_isInherited": false,
         "_autoScroll": true,
         "_scrollDuration": 500,
         "_onChildren": true,
